visuals: Use opacity
Experiment with shadows and opacity for other map elements.
I was going to add enemy shadows but it looks kinda ugly and
the shadow physics would need much more work to look remotely
ok. But now at least the wall shadows use opacity which simplifies
the code and makes it possible to cast the shadow on different
objects (though, as the perspective is kind of messed up in this
game anyway, it looks ugly if things that are not "laying on the ground"
are in the shadow).

Also, make level exits slightly opaque. This actually fixes the issue
where an enemy ends up in the exit tile by making it visible even in
that case.