Add a simple testing utility
Add a simple testing utility that starts the game with various
pre-recorded files and checks the game state after the game.
This is a rough smoke test to see if everything works like they should.
Test is compiled and run by running test.bat. The game must be compiled
separately.