Add xp and perks
Enemies give player experience points when killed. The amount of XP is
calculated from the enemy properties (similar to power score in random
dungeon generator script). XP can be spent to buy "perks" that are small
upgrades to player and player's skils. Perk buy menu is in the in-game
main menu and perks cannot be bought in arena fights.

Each perk will cost 2 times more than the previous one.

Game playback implementation is a bit of a hack: the 16 MSB bits of the
mask are used for the new perk mask if the perk bit is set (2048).

This has some effect on the game balance but the perks are not that
powerful, they just make the game a little less deadly. So, maybe it's
a good handicap for a new player that hasn't played the game for 10
years like me :D