Add dmac authentication
Add dmac-sha1 authentication to make sure the core level pack isn't
modified and the save games aren't forged as a way to make e.g.
speedrunning more competitive. It's not perfect but it's something.

I'm using a public domain SHA1 library github.com/clibs/sha1.
I chose this one because it was simple to integrate and the license
of course is very permissive (unlike many others that use e.g. GPL
licensing).

I also added a tool for automatically updating hashes for a level pack.