Reduce dependency on display rate
I discovered a major bug when running the game on lower display rates
such as 30 Hz. This was caused by the way the game uses the same
wait_event function for timing the game and handling IO events.
The display rate affects how fast the al_flip_display function returns
because it may wait for vsync. In the previous version the logic had a
couple of false assumptions:
- In menu, the screen is flipped and the logic is slowed by a 30 ms
  wait. When the display flip itself takes more time than expected
  the event queue never gets empty so inputs become so laggy it's
  unusable.
- In game the game loop has a wait function that calls the wait_event
  function only if the game clock and real clock have small enough
  difference. When the display flip itself takes more time than the
  "game tick rate", the wait_event function is never called.

 Therefore, this commit separates the timer queue from the IO queue.
 The IO queue is never waited, it's just emptied periodically.

 I experimented with frame skips to keep the frame rate consistent but
 it just looks janky and not worth it. So at lower display rates the
 main logic will just run slower.