Add mission pack "classic"
"Classic" is the game converted from old mission format, from files
dating back to 2011. The game tuning tries to be as close to the
experience back then, but there are some inevident differences.
The boss in particular is different as there's no way to allow player
to have infinite ammo just in that mission. The new boss is as
frustrating as the original boss but has a bit more variety.

Also, the turret powerup is a bit different. The original unloads its
shots slowly, where the current is very fast. It should be similar in
power, though.

The game is not possible in potion only mode.